feat: :lipstick: Configure app colors and add dark mode  (#20)
* feat(shared): :lipstick: Add and configure setup for dark mode

* feat: :lipstick: Set colors

* fix(spotify): :rotating_light: Fix lining warniing aboud end/start in flex

* feat: :lipstick: Set up dark/light mode for the app

* fix(spotify): :lipstick: Fix flier background

* docs: :memo: Change project name in package.json

* feat: :lipstick: Min color variables

* build: :green_heart: Fix redirects on netlify

* feat: :lipstick: add generated theme

* Add logo in selected color scheme

* feat(shared): :lipstick: Use color variables in all the styling files

+++ libs/shared/service/src/lib/user-settings.service.ts
@@ -17,7 +17,7 @@ export class UserSettings {
 export class UserSettingsService extends StateInterface<UserSettings> {
   protected readonly cookies_allowed = 'cookies_allowed';
   protected readonly dark_mode = 'dark_mode';
-  protected readonly return_path = 'dark_mode';
+  protected readonly return_path = 'return_path';
   protected readonly true_str = 'true';
 
   constructor(protected cookieService: CookieService) {
@@ -58,10 +58,11 @@ export class UserSettingsService extends StateInterface<UserSettings> {
   }
 
   getReturnPath(): string {
-    const path = this.cookieService.check(this.return_path)
-      ? this.cookieService.get(this.return_path)
-      : '';
-    this.cookieService.delete(this.return_path);
+    let path = '';
+    if (this.cookieService.check(this.return_path)) {
+      path = this.cookieService.get(this.return_path);
+      this.cookieService.delete(this.return_path);
+    }
     return path;
   }
 }
